Initial commit of working image link grabber
Right now it grabs and stores in a set a link for every image from the mars
images page.  It's not smart and requires a max number of pages, and it skips
image pages that do not have multiple resolutions -- but it works!

Now I need to actually download the damn images..